+++ ClientApp/src/components/News/News.js
@@ -3,12 +3,11 @@ import { Table } from 'reactstrap';
 import api from '../../sideEffects/apis/api'
 import { useForm, useFormFile } from '../../customHooks/useForm'
 import uploader from '../../customHooks/useFileUpload'
+import TextEditor from '../TextEditor/TextEditor';
 
 const noDataItem = {
     colSpan: '4',
-    style: {
-        textAlign: 'center'
-    }
+    
 }
 
 
@@ -21,7 +20,10 @@ const NewsList = ({ setNews, isBusy, ...props }) => {
                 setNews(news)
                 isBusy(false)
             })
-            .catch(err => console.error(err))
+            .catch(err => {
+                isBusy(false)
+                console.error(err)
+            })
     }, [setNews, isBusy])
 
     const createNewNews = (newsObject) => {
@@ -31,7 +33,11 @@ const NewsList = ({ setNews, isBusy, ...props }) => {
                 props.createNews(news)
                 isBusy(false)
             })
-            .catch(err => console.error(err))
+            .catch(err => {
+                isBusy(false)
+                console.error(err)
+            })
+        
     }
 
     const editNews = (newsObject) => {
@@ -41,30 +47,39 @@ const NewsList = ({ setNews, isBusy, ...props }) => {
                 props.editNews(news)
                 isBusy(false)
             })
-            .catch(err => console.error(err))
+            .catch(err => {
+                isBusy(false)
+                console.error(err)
+            })
     }
 
     const deleteNews = (id) => {
-        isBusy(true)
-        api.deleteWithId(`news/${id}`)
-            .then(news => {
-                props.deleteNews(id)
-                isBusy(false)
-            })
-            .catch(err => console.error(err))
+        if (window.confirm("Should we delete this?")) {
+            isBusy(true)
+            api.deleteWithId(`news/${id}`)
+                .then(news => {
+                    props.deleteNews(id)
+                    isBusy(false)
+                })
+                .catch(err => {
+                    isBusy(false)
+                    console.error(err)
+                })
+        }
+        
     }
 
 
     return (
-        <div>
+        <div className="f-width">
             <button type="button" onClick={() => setDisplay(!display)}>Add News</button>
+            <NewsForm className={display ? 'd-block' : 'd-none'} creator={true} isBusy={isBusy} onSubmit={createNewNews} />
             <Table striped>
                 <thead>
                     <tr>
                         <th>S/N</th>
                         <th>Title</th>
                         <th>Brief</th>
-                        <th>Date Posted</th>
                         <th>Actions</th>
                     </tr>
                 </thead>
@@ -80,7 +95,7 @@ const NewsList = ({ setNews, isBusy, ...props }) => {
                 }
 
             </Table>
-            <NewsForm className={display ? 'd-block' : 'd-none'} creator={true} isBusy={isBusy} onSubmit={createNewNews} />
+            
         </div>
         
     )
@@ -101,7 +116,6 @@ const News = (props) => {
                 <td>{props.sn}</td>
                 <td>{props.title}</td>
                 <td>{props.brief}</td>
-                <td>{new Date(props.datePosted).toString()}</td>
                 <td>
                     <button type='button' onClick={() => props.onDelete(props.id)}>
                         Delete
@@ -127,6 +141,7 @@ const NewsForm = (props) => {
     const contentField = useForm(null, (props.content || ''))
     const fileField = useFormFile()
     const filesUpload = uploader.useFileUpload();
+    //const [className, setClassName] = useState(props.className);
 
     const prepareData = () => {
         const preparatory = {
@@ -163,43 +178,47 @@ const NewsForm = (props) => {
     const fileUploadHandler = (e) => {
         props.isBusy(true)
         filesUpload.onUpload(e, fileField.value)
-        props.isBusy(false)
+            .then(() => props.isBusy(false))
+        
     }
 
     return (
-        <form className={props.className} onSubmit={submitForm}>
-            {props.image && <img src={filesUpload.value || props.image} alt={props.title} className='img-fluid' />}
-            <input {...idField.main} />
-            <div>
-                <label>Add Title: </label>
-                <input {...titleField.main} />
-            </div>
-            <div>
-                <label>Add Brief: </label>
-                <input {...briefField.main} />
-            </div>
-            <div>
-                <label>Add Content: </label>
-                <textarea {...contentField.main}></textarea>
-            </div>
-            <div>
-                <label>Upload Image: </label>
-                <input type='file' onChange={fileField.onChange} />
-            </div>
-            {fileField.value &&
-                <button type="button" onClick={fileUploadHandler}>
-                    {props.creator && 'Add news image'}
-                    {!props.creator && props.image && 'Edit image'}
-                    {!props.creator && !props.image && 'Add an image to this news'}
+        <div className="g-form">
+            <form className={props.className} onSubmit={submitForm}>
+                {props.image && <img src={filesUpload.value || props.image} alt={props.title} className='img-fluid' />}
+                <input {...idField.main} />
+                <div className="">
+                    <label>Add Title: </label>
+                    <input {...titleField.main} />
+                </div>
+                <div>
+                    <label>Add Brief: </label>
+                    <TextEditor {...briefField.main} />
+                </div>
+                <div>
+                    <label>Add Content: </label>
+                    <TextEditor {...contentField.main} />
+                </div>
+                <div>
+                    <label>Upload Image: </label>
+                    <input type='file' onChange={fileField.onChange} />
+                </div>
+                {fileField.value &&
+                    <button type="button" onClick={fileUploadHandler}>
+                        {props.creator && 'Add news image'}
+                        {!props.creator && props.image && 'Edit image'}
+                        {!props.creator && !props.image && 'Add an image to this news'}
+                    </button>
+                }
+
+                <br />
+                <button type='submit'>
+                    {props.creator ? 'Create News' : 'Edit news'}
                 </button>
-            }
-            
-            <br />
-            <button type='submit'>
-                {props.creator ? 'Create News' : 'Edit news'}
-            </button>
-            {!props.creator && <button onClick={(e) => props.onDelete(idField.main.value)} type='button'>Delete News</button>}
-        </form>
+                {!props.creator && <button onClick={(e) => props.onDelete(idField.main.value)} type='button'>Delete News</button>}
+            </form>
+        </div>
+       
 
     )
 }
